XlaEinsum should return unknown shape rather than fail on unranked inputs.
The op may exist in a function and is legal with unknown ranks. It is only when
refined at runtime that the shape needs to be static and known, representationally if it has unranked inputs it has unknown output shape.

This does not attempt to use the equation to determine ranks, that might be possible and could be refined to do so in future if deemed useful.

PiperOrigin-RevId: 254206933