Add DeviceAttributes to the collectives device resolver interface.
This change surfaces DeviceAttributes via the DeviceResolverInterface, by
changing GetDeviceLocalityAsync to GetDeviceAttributesAsync.  The eventual goal
is to pass device incarnation numbers of the remote device as a part of
collective ops's RPCs, and use them to detect worker failures.  We can still
access DeviceLocality as it is a part of DeviceAttributes.

As a part of this change, we also add the ability to clear the device resolver
cache upon a FailedPrecondition error.  The intention is to throw this error
upon detecting worker restart due to device incarnation mismatch.

PiperOrigin-RevId: 254784520