AveragePrecision@k(e) for an example e is defined as   Sum_{i=1,...,k} Precision@i(e) * I(i, e) / D(k, e) where I(i, e) is 1 if label of rank i for e is a positive label for e, and 0 otherwise, and the denominator D(k, e) is the maximum possible number of terms in the sum with a non-zero I(i, e): D(k, e) = min(k, number positive labels for e).
While this formula is implemented correctly in `average_precision_at_k` when `labels` is a sparse tensor, it is incorrect when `labels` is a dense tensor with some values out of range [0, num_classes), necessary in the case of examples with a different number of positive labels. In this case, the current implementation defines the denominator as
D(k, e) = min(k, size of last dimension of `labels`).

Consider 2 examples e_1 and e_2 with positive labels {0, 1} and {0}, assume that the top ranked label for e_2 is 0. AveragePrecision@2(e_2) = 1 and is computed correctly when the labels are given by a sparse tensor, but when the labels are given by a dense tensor, e.g. [[0, 1], [0, -1]], then AveragePrecision@2(e_2) = 0.5, as D(i, e_2) is considered min(2, 2).

This CL corrects this error.

PiperOrigin-RevId: 255283194