[XLA] Disambiguate reshape's dynamic output dimension when output is decomposed by input.
This is used as a hint for dynamic dimension inference to to know
which dimension is inferred by upper level framework. The problem it
tries to solve is this pattern that the pass sees:

  [<=20]       // 20 is dynamic
   |  Reshape
  [1, 20]    // which is dynamic?

Originally we say if the dimension 20 is unchanged, then in the result shape 20 is dynamic.

However, in some models we see this mirror pattern:

  [<=1, 20]    // 1 is dynamic
   |  Reshape
  [20]       // 20 is dynamic
   |  Reshape
  [<=1, 20]    // 1 is dynamic

As one can see, both "1" and "20" can be a valid dynamic dimension when reshaping from [20] to [1, 20]. It
is hard to tell which result dimension is dynamic without a hint.

There are three things we need in order to disambiguate:

1. A new "inferred_dimensino" field .
This CL pump through a new field "inferred dimension" through bridge
to HLO as a hint for the dynamic_dimension_inference pass to know
which dimension in the result is dynamic in reshape.

2. A simple constraint system that encodes how the dynamic dimension is generated in the first place.
See comments in dynamic_dimension_inference.h and dynamic_dimension_inference.cc

PiperOrigin-RevId: 255696421