[XLA] Merge broadcasts of effective scalar constants within fusions into their users in graphical HLO dump.
Just as we merge constants, merge broadcasts of effective scalar constants.

I don't merge non-scalar constants because then you'd have to specify a
broadcast dimension, and that gets complicated.

I only do this within fusions because outside of fusions a broadcast of a
constant is a "real" operation that you may want to see explicitly.

PiperOrigin-RevId: 256907877