[ROCm] Skipping subtests that check support for int8x4 type (on the GPU)
ROCm platform currently does not support int8x4 type (on the GPU)

This commit skips subtests (within python unit-tests) that test this functionality. The "skip" is guarded by the call to "is_built_with_rocm()", and hence these unit-tests will not be affected in any way when running with TF which was not built with ROCm support (i.e. `--config=rocm`)